Fix rtl placement for tooltip
